create vert tool (#49)
I think the new point tool is a good idea
RM had one put in place originally and we kept it in for awhile. Then the toolbar got out of hand and I tried to be too-clever-by-half in removing it. Now I think it's a real problem with people not knowing how to add a point
func create_tool_button(icon, tooltip) tidies things up nicely