Improve anchor positioning on non-bezier segments
Due to using cubic bezier, an anchor offset of 0.1 does not correspond
to 10% of the curve length, as one would expect. This effect is
especially visible when no bezier curves are used.
To mitigate this issue, shape_anchor now uses linear interpolation for
segments that don't make use of bezier curves.