Remove threaded edge generation because it is actually slower
This is unsurprising because spawning threads is slow.
There might be a sweet spot where a certain amount of index maps is
faster to process with a certain amount of threads.
But this requires more in-depth analysis.

The following measurements have been made with the test_corners.tscn
example which only has one index map per mesh.
"Update time" measures the whole duration of `force_update()`.

| Shape | Update time (with threads) | Update time (no threads) |
|-------|----------------------------|--------------------------|
| Left  | 4.189 ms                   | 3.297 ms                 |
| Right | 2.118 ms                   | 1.694 ms                 |