Rework Collision Tool
The previous behavior created a StaticBody2D and an invisible child
CollisionPolygon2D, then also moved the shape to be a child of the
StaticBody2D.

StaticBody2D
|- CollisionPolygon
|- Shape

This behavior was pretty useless in real world projects and caused
confusion among new users (Godot and/or SS2D).

- People thought this hierarchy is mandatory and were irritated how
  inflexible it is
- In larger projects you often do not use bare StaticBodies but
  reusable scenes that store more information about the body, e.g.
  collision layers, terrain type, additional nodes, etc.
- It might be desirable to put multiple colliders into the same
  StaticBody or not to use StaticBody2D at all but CharacterBody2D or
  Area2D instead.

For this reason, the collision tool was reworked to generate only a
CollisionPolygon2D and add it as sibling.
This behavior is more straight forward and grants users the flexibility
to use it however they want without having to partially undo the
generated hierarchy (or being confused about it).

Additionally, a "Collision Generation Options" dialog was added where
the parent for newly created CollisionPolygons can be set.
If unset or not found, the node is added as sibling to the shape.
The path can be specified as a regular node path or as group name.
This feature is useful to group all Colliders under a single
StaticBody2D without moving it manually.

Finally, this patch introduces SS2D_EditorTool as a base class for
editor tools with the goal of moving tool code out of plugin.gd.
The Collision Tool is now implemented in SS2D_CollisionEditorTool.