Prevent crash when modifying the shape before _ready()
Shapes becomes dirty upon scene load. The dirty update is called
deferred, so `_ready()` already ran.
This does not seem to be the case when calling
`get_tree().reload_current_scene()`. For whatever reason, the deferred
`_on_dirty_update` call already runs before the node was added to the
scene.

This patch prevents dirty updates before the shape is ready.

Closes #197